+++ packages/editor/editor-core/src/plugins/table/utils/colgroup.ts
@@ -5,12 +5,18 @@ export const generateColgroup = (node: PmNode) => {
   const cols: Array<string | { [attr: string]: string } | {}> = [];
 
   node.content.firstChild!.content.forEach(cell => {
+    const colspan = cell.attrs.colspan || 1;
     if (Array.isArray(cell.attrs.colwidth)) {
-      cell.attrs.colwidth.forEach(width => {
+      // We slice here to guard against our colwidth array having more entries
+      // Than the we actually span. We'll patch the document at a later point.
+      cell.attrs.colwidth.slice(0, colspan).forEach(width => {
         cols.push(['col', { style: `width: ${width}px;` }]);
       });
     } else {
-      cols.push(['col', {}]);
+      // When we have merged cells on the first row (firstChild),
+      // We want to ensure we're creating the appropriate amount of
+      // cols the table still has.
+      cols.push(...Array.from({ length: colspan }, _ => ['col', {}]));
     }
   });