+++ packages/editor/renderer/src/__tests__/visual-regression/layouts.ts
@@ -1,4 +1,5 @@
 import { goToRendererTestingExample, mountRenderer, snapshot } from './_utils';
+import { layoutWithDefaultBreakoutMark } from './__fixtures__/document-with-layout-default-breakout';
 
 const twoColumnLayout = {
   version: 1,
@@ -66,5 +67,25 @@ describe.skip('Snapshot Test: Layouts', () => {
       });
       await snapshot(page);
     });
+  });
+});
+
+describe('Snapshot Test: Breakout Layouts', () => {
+  let page;
+  beforeAll(async () => {
+    // @ts-ignore
+    page = global.page;
+    await goToRendererTestingExample(page);
+  });
+
+  it(`should correctly render two column layout with a default breakout mark`, async () => {
+    await page.setViewport({ width: 1120, height: 500 });
+    await page.waitFor(100);
+    mountRenderer(page, {
+      document: layoutWithDefaultBreakoutMark,
+      appearance: 'full-page',
+    });
+    await page.waitFor(100);
+    await snapshot(page);
   });
 });