+++ packages/elements/user-picker/src/components/MultiValue.tsx
@@ -1,10 +1,13 @@
 import Tag from '@atlaskit/tag';
 import * as React from 'react';
 import { FormattedMessage } from 'react-intl';
-import { Option } from '../types';
+import { Option, UserPickerProps } from '../types';
+import { AddOptionAvatar } from './AddOptionAvatar';
 import { messages } from './i18n';
 import { SizeableAvatar } from './SizeableAvatar';
-import { getAvatarUrl } from './utils';
+import { getAvatarUrl, isEmail } from './utils';
+import styled from 'styled-components';
+import { colors } from '@atlaskit/theme';
 
 export const scrollToValue = (
   valueContainer: HTMLDivElement,
@@ -20,12 +23,20 @@ export const scrollToValue = (
   }
 };
 
+const TagContainer = styled.div`
+  button {
+    &:focus {
+      box-shadow: 0 0 0 2px ${colors.B100};
+    }
+  }
+`;
+
 type Props = {
   isFocused?: boolean;
   data: Option;
   innerProps: any;
   removeProps: { onClick: Function };
-  selectProps: { isDisabled: boolean };
+  selectProps: UserPickerProps;
 };
 
 export class MultiValue extends React.Component<Props> {
@@ -73,6 +84,29 @@ export class MultiValue extends React.Component<Props> {
     );
   }
 
+  getElemBefore = () => {
+    const {
+      data: { data, label },
+      selectProps,
+    } = this.props;
+    if (isEmail(data)) {
+      return selectProps.emailLabel ? (
+        <AddOptionAvatar size="small" label={selectProps.emailLabel} />
+      ) : (
+        <FormattedMessage {...messages.addEmail}>
+          {label => <AddOptionAvatar size="small" label={label as string} />}
+        </FormattedMessage>
+      );
+    }
+    return (
+      <SizeableAvatar
+        appearance="multi"
+        src={getAvatarUrl(data)}
+        name={label}
+      />
+    );
+  };
+
   render() {
     const {
       data: { label, data },
@@ -86,21 +120,17 @@ export class MultiValue extends React.Component<Props> {
       <div ref={this.containerRef}>
         <FormattedMessage {...messages.remove}>
           {remove => (
-            <Tag
-              {...innerProps}
-              appearance="rounded"
-              text={label}
-              elemBefore={
-                <SizeableAvatar
-                  appearance="multi"
-                  src={getAvatarUrl(data)}
-                  name={label}
-                />
-              }
-              removeButtonText={data.fixed || isDisabled ? undefined : remove}
-              onAfterRemoveAction={onRemove}
-              color={isFocused ? 'blueLight' : undefined}
-            />
+            <TagContainer>
+              <Tag
+                {...innerProps}
+                appearance="rounded"
+                text={label}
+                elemBefore={this.getElemBefore()}
+                removeButtonText={data.fixed || isDisabled ? undefined : remove}
+                onAfterRemoveAction={onRemove}
+                color={isFocused ? 'blueLight' : undefined}
+              />
+            </TagContainer>
           )}
         </FormattedMessage>
       </div>