+++ packages/elements/user-picker/src/components/Option.tsx
@@ -1,23 +1,42 @@
 import { components } from '@atlaskit/select';
 import * as React from 'react';
-import { UserOption } from './UserOption';
+import { Option as OptionType, UserPickerProps } from '../types';
+import { EmailOption } from './EmailOption';
 import { TeamOption } from './TeamOption';
-import { isUser, isTeam } from './utils';
+import { UserOption } from './UserOption';
+import { isEmail, isTeam, isUser } from './utils';
 
-export const Option = (props: any) => {
-  const {
-    data: { data },
-    status,
-    isSelected,
-  } = props;
-
-  return (
-    <components.Option {...props}>
-      {isUser(data) ? (
-        <UserOption user={data} status={status} isSelected={isSelected} />
-      ) : (
-        isTeam(data) && <TeamOption team={data} isSelected={isSelected} />
-      )}
-    </components.Option>
-  );
+export type OptionProps = {
+  data: OptionType;
+  isSelected: boolean;
+  status?: string;
+  selectProps: UserPickerProps;
 };
+
+const dataOption = ({
+  data: { data },
+  isSelected,
+  status,
+  selectProps,
+}: OptionProps) => {
+  if (isUser(data)) {
+    return <UserOption user={data} status={status} isSelected={isSelected} />;
+  }
+  if (isEmail(data)) {
+    return (
+      <EmailOption
+        email={data}
+        isSelected={isSelected}
+        label={selectProps.emailLabel}
+      />
+    );
+  }
+  if (isTeam(data)) {
+    return <TeamOption team={data} isSelected={isSelected} />;
+  }
+  return null;
+};
+
+export const Option: React.StatelessComponent<OptionProps> = props => (
+  <components.Option {...props}>{dataOption(props)}</components.Option>
+);