+++ packages/media/media-card/src/styles/mixins.tsx
@@ -1,5 +1,5 @@
 /* tslint:disable:variable-name */
-import { akBorderRadius } from '@atlaskit/util-shared-styles';
+import { borderRadius } from '@atlaskit/theme';
 import { CardAppearance } from '../index';
 import { defaultTransitionDuration } from './config';
 
@@ -33,8 +33,8 @@ export const centerSelf = `
 `;
 
 export const borderRadiusLeft = `
-  border-top-left-radius: ${akBorderRadius};
-  border-bottom-left-radius: ${akBorderRadius};
+  border-top-left-radius: ${borderRadius()};
+  border-bottom-left-radius: ${borderRadius()};
 `;
 
 export const spaceAround = `