+++ packages/media/media-editor/src/react/editorView/toolbar/popups/colorButtonStyles.tsx
@@ -3,16 +3,16 @@
 import styled from 'styled-components';
 
 import { HTMLAttributes, ComponentClass } from 'react';
-import { akColorN0, akBorderRadius } from '@atlaskit/util-shared-styles';
+import { colors, borderRadius } from '@atlaskit/theme';
 
 export const ColorSample: ComponentClass<HTMLAttributes<{}>> = styled.div`
   cursor: pointer;
   width: 24px;
   height: 24px;
   margin: 4px;
-  border-radius: ${akBorderRadius};
+  border-radius: ${borderRadius()};
 `;
 
 export const CheckArea: ComponentClass<HTMLAttributes<{}>> = styled.div`
-  color: ${akColorN0};
+  color: ${colors.N0};
 `;