+++ packages/media/media-viewer/src/newgen/media-viewer.tsx
@@ -3,24 +3,46 @@ import { Context } from '@atlaskit/media-core';
 import { IntlProvider, intlShape } from 'react-intl';
 import { ThemeProvider } from 'styled-components';
 import { Shortcut, theme } from '@atlaskit/media-ui';
+import { withAnalyticsEvents } from '@atlaskit/analytics-next';
+import { WithAnalyticsEventProps } from '@atlaskit/analytics-next-types';
+import { mediaViewerModalEvent } from './analytics/media-viewer';
+import { channel } from './analytics/index';
+import {
+  GasPayload,
+  GasScreenEventPayload,
+} from '@atlaskit/analytics-gas-types';
 import { Identifier, ItemSource, MediaViewerFeatureFlags } from './domain';
 import { List } from './list';
 import { Collection } from './collection';
 import { Content } from './content';
 import { Blanket } from './styled';
 
-export type Props = Readonly<{
-  onClose?: () => void;
-  selectedItem?: Identifier;
-  featureFlags?: MediaViewerFeatureFlags;
-  context: Context;
-  itemSource: ItemSource;
-}>;
+export type Props = Readonly<
+  {
+    onClose?: () => void;
+    selectedItem?: Identifier;
+    featureFlags?: MediaViewerFeatureFlags;
+    context: Context;
+    itemSource: ItemSource;
+  } & WithAnalyticsEventProps
+>;
 
-export class MediaViewer extends React.Component<Props, {}> {
+class MediaViewerComponent extends React.Component<Props, {}> {
   static contextTypes = {
     intl: intlShape,
   };
+
+  private fireAnalytics = (payload: GasPayload | GasScreenEventPayload) => {
+    const { createAnalyticsEvent } = this.props;
+    if (createAnalyticsEvent) {
+      const ev = createAnalyticsEvent(payload);
+      ev.fire(channel);
+    }
+  };
+
+  componentWillMount() {
+    this.fireAnalytics(mediaViewerModalEvent());
+  }
 
   render() {
     const { onClose } = this.props;
@@ -74,3 +96,5 @@ export class MediaViewer extends React.Component<Props, {}> {
     }
   }
 }
+
+export const MediaViewer = withAnalyticsEvents()(MediaViewerComponent);